Added form Validation

+++ SoundPacker/SoundPackingLib.cpp
@@ -4,7 +4,7 @@ using namespace std;
 using namespace System;
 using namespace System::Diagnostics;
 
-int runAlgorithm(string inputFile, string algorithm, float duration, bool allowCopy) {
+double runAlgorithm(string inputFile, string algorithm, float duration, bool allowCopy) {
 	
 	return 0;
 }
